update to always look for quality 265

+++ radarr_unmonitor.ps1
@@ -1,8 +1,8 @@
 $expected_eventtype = 'Download'
 $search_text_1 = 'lama'
 $search_text_2 = 'rarbg'
-$search__sub_text_2 = '265'
 $search_text_3 = 'psa'
+$search__quality = '265'
 $doApiCall = $false;
 
 #$env:radarr_eventtype = 'Download'
@@ -14,18 +14,18 @@ $doApiCall = $false;
 if ($env:radarr_eventtype -eq $expected_eventtype) {
   ('{0} - In expected if statement') -f (Get-Date) | Out-File -FilePath C:\radarr_unmonitor.txt -Append
 
-  if ($env:radarr_moviefile_relativepath -like ('*{0}*' -f [WildcardPattern]::Escape($search_text_1))) {
+  if ($env:radarr_moviefile_relativepath -like ('*{0}*' -f [WildcardPattern]::Escape($search_text_1)) -And $env:radarr_moviefile_relativepath -like ('*{0}*' -f [WildcardPattern]::Escape($search__quality))) {
     ('{0} - Search text 1 ''' + $search_text_1 + ''' found') -f (Get-Date) | Out-File -FilePath C:\radarr_unmonitor.txt -Append
     $doApiCall = $true;
   }
 
-  if ($env:radarr_moviefile_relativepath -like ('*{0}*' -f [WildcardPattern]::Escape($search_text_3))) {
-    ('{0} - Search text 1 ''' + $search_text_3 + ''' found') -f (Get-Date) | Out-File -FilePath C:\radarr_unmonitor.txt -Append
+  if ($env:radarr_moviefile_relativepath -like ('*{0}*' -f [WildcardPattern]::Escape($search_text_2)) -And $env:radarr_moviefile_relativepath -like ('*{0}*' -f [WildcardPattern]::Escape($search__quality))) {
+    ('{0} - Search text 2 ''' + $search_text_2 + ''' found') -f (Get-Date) | Out-File -FilePath C:\radarr_unmonitor.txt -Append
     $doApiCall = $true;
   }
 
-  if ($env:radarr_moviefile_relativepath -like ('*{0}*' -f [WildcardPattern]::Escape($search_text_2)) -And $env:radarr_moviefile_relativepath -like ('*{0}*' -f [WildcardPattern]::Escape($search__sub_text_2))) {
-    ('{0} - Search text 2 ''' + $search_text_2 + ''' found') -f (Get-Date) | Out-File -FilePath C:\radarr_unmonitor.txt -Append
+  if ($env:radarr_moviefile_relativepath -like ('*{0}*' -f [WildcardPattern]::Escape($search_text_3)) -And $env:radarr_moviefile_relativepath -like ('*{0}*' -f [WildcardPattern]::Escape($search__quality))) {
+    ('{0} - Search text 1 ''' + $search_text_3 + ''' found') -f (Get-Date) | Out-File -FilePath C:\radarr_unmonitor.txt -Append
     $doApiCall = $true;
   }
